Move the targets and targetsFinder to its own services folder
