+++ src/services/configurations/index.js
@@ -1,11 +1,9 @@
 const { babelConfiguration } = require('./babelConfiguration');
 const { projectConfiguration } = require('./projectConfiguration');
 const { targetConfiguration } = require('./targetConfiguration');
-const { targetsFinder } = require('./targetsFinder');
 
 module.exports = {
   babelConfiguration,
   projectConfiguration,
   targetConfiguration,
-  targetsFinder,
 };
