+++ src/services/targets/index.js
@@ -1,0 +1,7 @@
+const { targets } = require('./targets');
+const { targetsFinder } = require('./targetsFinder');
+
+module.exports = {
+  targets,
+  targetsFinder,
+};
